fixed black image
