+++ invokeai/frontend/web/src/app/store/middleware/listenerMiddleware/listeners/enqueueRequestedLinear.ts
@@ -24,7 +24,6 @@ export const addEnqueueRequestedLinear = (startAppListening: AppStartListening) 
         graph = await buildGenerationTabGraph(state);
       }
 
-      console.log("Here's the graph:", graph);
       const batchConfig = prepareLinearUIBatch(state, graph, prepend);
 
       const req = dispatch(
